Player animations and background uploaded to the Unity project
I uploaded the player animations/sprite sheets into a folder called "Player Art" in the main Assets folder. I also uploaded the background for level 1 in a folder called "Background" in the main Assets folder. The background has been added to the scene as well and scaled up to 4.16 to make it fit the size of the scene. The sides go outside of the screen as well.