Added audio
Added audio files for background music, ambient sounds, and various actions. Added an Audio Manager prefab with script attached. Background music and ambient sounds should automatically play when the scene is started. Modified the Player Controller script to play the "jump" and "hit" noises when the player jumps or hits an enemy.